+++ app/src/main/java/ps/reso/instaeclipse/mods/ui/UIHookManager.java
@@ -1,0 +1,192 @@
+package ps.reso.instaeclipse.mods.ui;
+
+import static ps.reso.instaeclipse.mods.ghost.ui.GhostEmojiManager.addGhostEmojiNextToInbox;
+
+import android.annotation.SuppressLint;
+import android.app.Activity;
+import android.os.Handler;
+import android.os.Looper;
+import android.util.Log;
+import android.view.View;
+import android.view.ViewGroup;
+import android.widget.Toast;
+
+import de.robv.android.xposed.XC_MethodHook;
+import de.robv.android.xposed.XposedBridge;
+import de.robv.android.xposed.XposedHelpers;
+import ps.reso.instaeclipse.mods.devops.config.ConfigManager;
+import ps.reso.instaeclipse.mods.ui.utils.VibrationUtil;
+import ps.reso.instaeclipse.utils.dialog.DialogUtils;
+import ps.reso.instaeclipse.utils.feature.FeatureFlags;
+import ps.reso.instaeclipse.utils.ghost.GhostModeUtils;
+
+public class UIHookManager {
+
+    @SuppressLint("StaticFieldLeak")
+    private static Activity currentActivity;
+
+    public static Activity getCurrentActivity() {
+        return currentActivity;
+    }
+    private static boolean isAnyGhostOptionEnabled() {
+        return GhostModeUtils.isGhostModeActive();
+    }
+
+    public void mainActivity(ClassLoader classLoader) {
+        // Hook onCreate of Instagram Main
+        XposedHelpers.findAndHookMethod("com.instagram.mainactivity.InstagramMainActivity", classLoader, "onCreate", android.os.Bundle.class, new XC_MethodHook() {
+            @Override
+            protected void afterHookedMethod(MethodHookParam param) {
+                final Activity activity = (Activity) param.thisObject;
+                currentActivity = activity;
+                activity.runOnUiThread(() -> {
+                    try {
+                        setupHooks(activity);
+                        addGhostEmojiNextToInbox(activity, isAnyGhostOptionEnabled());
+                    } catch (Exception ignored) {
+
+                    }
+                });
+            }
+        });
+
+        // Hook onResume - Instagram Main
+        XposedHelpers.findAndHookMethod("com.instagram.mainactivity.InstagramMainActivity", classLoader, "onResume", new XC_MethodHook() {
+            @Override
+            protected void afterHookedMethod(MethodHookParam param) {
+                final Activity activity = (Activity) param.thisObject;
+                currentActivity = activity;
+                activity.runOnUiThread(() -> {
+                    try {
+                        setupHooks(activity);
+                        addGhostEmojiNextToInbox(activity, isAnyGhostOptionEnabled());
+                        if (FeatureFlags.isImportingConfig) {
+                            FeatureFlags.isImportingConfig = false;
+                            ConfigManager.importConfigFromClipboard(activity);
+                        }
+
+                    } catch (Exception ignored) {
+                    }
+                });
+            }
+        });
+
+        // Hook getBottomSheetNavigator - Instagram Main
+        XposedHelpers.findAndHookMethod("com.instagram.mainactivity.InstagramMainActivity", classLoader, "getBottomSheetNavigator", new XC_MethodHook() {
+            @Override
+            protected void afterHookedMethod(MethodHookParam param) {
+                final Activity activity = getCurrentActivity();
+                if (activity != null) {
+                    activity.runOnUiThread(() -> {
+                        try {
+                            setupHooks(activity);
+                            addGhostEmojiNextToInbox(activity, GhostModeUtils.isGhostModeActive());
+                        } catch (Exception ignored) {
+                        }
+                    });
+                }
+            }
+        });
+
+        // Hook onResume - Model
+        XposedHelpers.findAndHookMethod("com.instagram.modal.ModalActivity", classLoader, "onResume", new XC_MethodHook() {
+            @Override
+            protected void afterHookedMethod(MethodHookParam param) {
+                Activity activity = (Activity) param.thisObject;
+                if (activity != null) {
+                    activity.runOnUiThread(() -> {
+                        try {
+                            setupHooks(activity);
+                        } catch (Exception ignored) {
+                        }
+                    });
+                }
+            }
+        });
+    }
+
+    private static void setupHooks(Activity activity) {
+        // Hook Search Tab (open InstaEclipse Settings)
+        hookLongPress(activity, "search_tab", v -> {
+            DialogUtils.showEclipseOptionsDialog(activity);
+            VibrationUtil.vibrate(activity);
+            return true;
+        });
+
+        // Hook Inbox Button (toggle Ghost Quick Options)
+        String[] possibleIds = {"action_bar_inbox_button", "direct_tab"};
+
+        for (String id : possibleIds) {
+            @SuppressLint("DiscouragedApi") int viewId = activity.getResources().getIdentifier(id, "id", activity.getPackageName());
+            View view = activity.findViewById(viewId);
+            if (view != null) {
+                hookLongPress(activity, id, v -> {
+                    GhostModeUtils.toggleSelectedGhostOptions(activity);
+                    VibrationUtil.vibrate(activity);
+                    return true;
+                });
+                break;
+            }
+        }
+
+        addGhostEmojiNextToInbox(activity, GhostModeUtils.isGhostModeActive());
+
+        // Mark messages (DM) as seen by holding on gallery button
+        hookLongPress(activity, "row_thread_composer_button_gallery", v -> {
+            VibrationUtil.vibrate(activity);
+
+            if (!GhostModeUtils.isGhostModeActive()) {
+                return true;
+            }
+
+            FeatureFlags.isGhostSeen = false;
+
+            activity.getWindow().getDecorView().post(() -> {
+                try {
+                    // Look for the exact message list view by ID
+                    @SuppressLint("DiscouragedApi") int messageListId = activity.getResources().getIdentifier("message_list", "id", activity.getPackageName());
+                    View view = activity.findViewById(messageListId);
+
+                    if (view instanceof ViewGroup messageList) {
+
+                        // Try scrolling via translation if standard scroll methods don't exist
+                        messageList.scrollBy(0, -100); // scroll up
+
+                        new Handler(Looper.getMainLooper()).postDelayed(() -> {
+                            messageList.scrollBy(0, 100); // scroll back down
+
+                            FeatureFlags.isGhostSeen = true;
+                            Toast.makeText(activity, "✅ Message was marked as read", Toast.LENGTH_SHORT).show();
+
+                        }, 300);
+
+
+                    } else {
+                        XposedBridge.log("⚠️ message_list not a ViewGroup or not found — fallback to reset flag");
+
+                        new Handler(Looper.getMainLooper()).postDelayed(() -> FeatureFlags.isGhostSeen = true, 300);
+                    }
+                } catch (Exception e) {
+                    XposedBridge.log("❌ Exception in scroll logic: " + Log.getStackTraceString(e));
+                }
+            });
+
+            return true;
+        });
+
+    }
+
+    // Hook long press method
+    private static void hookLongPress(Activity activity, String viewName, View.OnLongClickListener listener) {
+        try {
+            @SuppressLint("DiscouragedApi") int viewId = activity.getResources().getIdentifier(viewName, "id", activity.getPackageName());
+            View view = activity.findViewById(viewId);
+
+            if (view != null) {
+                view.setOnLongClickListener(listener);
+            }
+        } catch (Exception ignored) {
+        }
+    }
+
+}
